# iteh code/eventapi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventHub API",
        "description": "API dokumentacija za EventHub aplikaciju",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registracija korisnika",
                "operationId": "278f3b96a2ff88d120714549e2113f4d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "example": "Stefan"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Malbasa"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "stefan@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uspešna registracija"
                    },
                    "422": {
                        "description": "Greška validacije"
                    }
                },
                "servers": [
                    {
                        "url": "http://localhost:8000",
                        "description": "Local server"
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Prijava korisnika",
                "operationId": "c6e01b33023a9a4b02f91c4fc666fe34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "stefan@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešna prijava"
                    },
                    "401": {
                        "description": "Neispravni kredencijali"
                    },
                    "422": {
                        "description": "Greška validacije"
                    }
                },
                "servers": [
                    {
                        "url": "http://localhost:8000",
                        "description": "Local server"
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Odjava korisnika",
                "operationId": "0a5c63b873b73047b8636bae7a0a2fcc",
                "responses": {
                    "200": {
                        "description": "Uspešna odjava"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "http://localhost:8000",
                        "description": "Local server"
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Podaci o trenutno ulogovanom korisniku",
                "operationId": "2c13fd98adec6f071980de5a39d14f49",
                "responses": {
                    "200": {
                        "description": "Podaci o korisniku"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "http://localhost:8000",
                        "description": "Local server"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}